requirements.txt
attrs==19.3.0
coverage==5.1
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
pytest-cov==2.8.1
python-decouple==3.3
six==1.14.0
wcwidth==0.1.9
zipp==3.1.0
